1723: Support grid 0.5, 0.25, 0.125 (#2116)
* 1723: Initial setup for sub-integer grid

* 2069: add const

* 2069: refactor grid index calculations

* 1723: add some tests. moveDeltaForFace_SubInteger is failing

* 1723: tweak gridSkip calculation

+++ common/src/Model/Brush.h
@@ -171,8 +171,8 @@ namespace TrenchBroom {
             bool canRemoveVertices(const BBox3& worldBounds, const Vec3::List& vertexPositions) const;
             void removeVertices(const BBox3& worldBounds, const Vec3::List& vertexPositions);
             
-            bool canSnapVertices(const BBox3& worldBounds, size_t snapTo);
-            void snapVertices(const BBox3& worldBounds, size_t snapTo);
+            bool canSnapVertices(const BBox3& worldBounds, FloatType snapTo);
+            void snapVertices(const BBox3& worldBounds, FloatType snapTo);
 
             // edge operations
             bool canMoveEdges(const BBox3& worldBounds, const Edge3::List& edgePositions, const Vec3& delta) const;
